requirements.txt
# 数据层
tushare>=1.2.80
akshare>=1.8.95
pandas>=2.0.0
numpy>=1.24.0

# 特征工程
ta-lib>=0.4.28
snownlp>=0.12.3
scikit-learn>=1.2.0
statsmodels>=0.14.0
seaborn>=0.12.0

# 预测模型
tensorflow>=2.12.0
# tensorflow-addons>=0.20.0  # 当前Python版本可能不兼容，需要时再安装
optuna>=3.0.0  # 用于超参数调优

# 回测框架
# backtrader>=1.9.76.123  # 项目已实现自定义回测引擎
# pyfolio>=0.9.2  # 项目已实现自定义绩效分析工具

# 交易接口
easytrader>=0.23.7  # 最新可用版本

# 可视化
plotly>=5.15.0
dash>=2.10.0
dash-bootstrap-components>=1.4.0
matplotlib>=3.7.0
seaborn>=0.12.0

# 数据库
# pysqlite3>=0.5.0  # Python 3.12标准库已包含sqlite3模块

# 科学计算
scipy>=1.10.0

# 并行计算
joblib>=1.2.0

# 工具库
# logging  # Python标准库，不需要额外安装
matplotlib>=3.7.0
schedule>=1.2.0
tqdm>=4.65.0
PyYAML>=6.0  # 用于YAML配置文件处理